+++ src/index.ts
@@ -1,9 +1,9 @@
-import { isDefined, isString, isNumber, isBigInt, isBool, isSymbol, isNull, } from './guards' /* prettier-ignore */
+import { isDefined, isString, isNumber, isBigInt, isBool, isSymbol, isNull, isObject } from './guards' /* prettier-ignore */
 import { generateError } from './utils'
 
 import type { JsonValue, JSONValue } from './JSON'
 import type { Tuple } from './tuple'
 import type { PropsOf } from './vue'
 
-export { isDefined, isString, isNumber, isBigInt, isBool, isSymbol, isNull, generateError } /* prettier-ignore */
+export { isDefined, isString, isNumber, isBigInt, isBool, isSymbol, isNull, generateError, isObject } /* prettier-ignore */
 export type { JsonValue, JSONValue, Tuple, PropsOf }