+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `JsonValue` in favor of `JSONValue`. Casing should be about word spacing only.
   Using all capitals for an abbreviation makes names harder to read.
+- `isObject(x)` - Is x an object?
 
 ### Deprecated